requirements.txt
flask==2.3.3
opencv-python==4.8.0.76
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
requests==2.31.0
tensorflow==2.12.0
ultralytics==8.0.0
matplotlib==3.7.1
pandas==2.0.3
scipy==1.10.1 